lib/thread: Add exit function parameter to spdk_thread_lib_init()
Even if a thread calls spdk_thread_exit(), and spdk_thread_destroy(),
its context is still registered in a reactor.  Hence when a thread
calls spdk_thread_exit() and spdk_thread_destroy() voluntarily,
application shutdown does not complete correctly.

This patch adds a parameter exit_thread_fn to spdk_thread_lib_init().
A function to remove the exited thread from the reactor is expected
to be set to the parameter and the change will be done in the next
patch.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I4736c91b5ffac9113d59a7388a39be350f702114