Update patch set 2
Patch Set 2:

Good point, I missed that as fc is not using listen/stop_listen from transport layer.

I'm not familiar with fc so any feedback or idea how to resolve is welcome. However, the alignment to changes is proposed with patchset #2. Let me know if it is acceptable.

Patch-set: 2