lib/blob: Fix free cluster count after power failure
At one point of snapshot creation we have two blobs with completely valid md,
same clusters and extent pages. The only indication that one of them is invalid,
is presented as xattr SNAPSHOT_IN_PROGRESS.
Flow here is that when we open the blob which has this xattr
we to open it and in _spdk_bs_examine_clone() check with its parent for correct id.
Then when they differ – meaning parent has conflicting metadata – we delete the corrupted blob.
In _spdk_bs_delete_corrupted_blob() each cluster is set to 0
and that corrupted blob is set to thin provisioned.

During _spdk_bs_load_replay_md_parse_page() bs->num_free_clusters
was decreased for each of the clusters read from md.

As those clusters will not be cleaned in usual path,
increase the num_free_clusters for each allocated cluster.

Changes in UT expose the issue and would not pass if
not for the fix in blobstore.c.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I8b28573cdc228c467747e619ee785700eb151a34