env_dpdk: allow multiple virt->phys mappings.
In vhost, when we have multiple devices registered and the iommu is
operating in pa mode, all of those devices try to map the same buffers
to different virtual addresses in the hosts memory space. This leads to
us calling the VFIO_IOMMU_MAP_DMA multiple times for the same physical
address which is not allowed. This change allows us to keep the multiple
virtual to physical mappings stored in the spdk memory map even when we
are unable to remap the vfio iommu mappings.

Change-Id: Id0eecb2ef0aa4fd51d383c946f104dc374cb0b41
Signed-off-by: Seth Howell <seth.howell@intel.com>