Update patch set 4
Patch Set 4:

this is a pretty small patch so I didn't break it up but in order to enable an RPC that takes the base bdev as input I had to add a global list of gpt vdbevs and their associated base bdevs that I could use to hunt down the right one (the trello card for this metions a global list but one did not exist). I can bust that out into another patch if anyone would like me to. I didn't add UT as this is primarily a test tool but I did test manually on an nvmedisk to confirm it behaves as the commit messages says it does.

Patch-set: 4