Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> I'd rather fix this in the posix implementation. I updated the patch before this one (https://review.gerrithub.io/c/spdk/spdk/+/483311) with what I think is the fix. We'll see if it passes.

Hi  Ben,

It is OK. Generally I think that there are two issues:
1 The socket can be closed in posix layer.
2 The socket can be removed from the group in the upper layer.

So It is ok that you merge the two patches into one in posix layer. For this patch, I put in geeneric layer since we may possibly face this issue later in other socket layer. But I think that you can put it both in posix layer to fix the issue.

Patch-set: 2