Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> I think this unnecessarily complicates the SPDK build.
> 
> It makes sense for SPDK to build its built-in dependencies, but if user wants to provide a custom version of e.g. DPDK it should be expected of him to build this version by himself.
> 
> $ ./configure --with-dpdk=[..] --force-dpdk-build
> 
> vs
> 
> $ cd /dpdk; ./configure --something, make
> $ cd /spdk; ./configure --with-dpdk=[...]

This change is for the case where user wants to configure DPDK the same way as if it with our submodule. Such as our vs-dpdk jobs where we separately configure DPDK trying to match how its done with submodule.
So without the --force-dpdk-build and the following patch, user needs to investigate on his own how we do DPDK build, adopt it to his version and maintain their build during updates.
`--something` is what makes second option harder to use.

Patch-set: 3