module/crypto: add RPC and plumbing to support QAT AES_XTS cipher
Did not add support for legacy config file use with AES_XTS. For
RPC support only the plumbing is in the patch, the rest is coming
up. The existing args for our python scripts were positional. The
two new args to create a vbdev using AES_XTS are optional which
appear to require them to be named so there's a mix now, not sure
how everyone feels about that.

Signed-off-by: paul luse <paul.e.luse@intel.com>
Change-Id: Iadd03f8da305e97e3f1d57ef4aeaece400e8d959