FIO_plugin: signal the condition at the cleanup phase
This is to solve below Github issue:
https://github.com/spdk/spdk/issues/1146

There are two related threads here. One is the init thread
which handles the initialization and the other is the fio
thread which handles the devices. The problem here is that
the close devices operation may be called in the init thread.

Fix here it to make sure the close devices operation in the
fio thread only.

When FIO stops, it will first cleanup then close the engine.
Signal the condition earlier at the cleanup phase so that
the init thread will be unlocked earlier to have cleanup work.

What's more, the fio thread which open the device will do the
close device only.

Change-Id: Icaacbb037233d600d4de9a2a6f39fe8f75562a79
Signed-off-by: GangCao <gang.cao@intel.com>