lib/event: Control CPU affinity of SPDK thread dynamically
Add an new API spdk_reactor_set_thread_affinity() to set the cpumask
of the thread to the specified value. If the thread is not currently
running on the specified CPUs, the thread is migrated to one of the
specified CPUs.

Details are as follows.

Add lcore to struct spdk_lw_thread to invoke the event to change the
CPU affinity of the thread on the CPU core which the thread runs on.

In the event handler _set_thread_affinity(), set the cpumask of the
thread to the specified value. If the thread is not running on one
of the specified CPUs, remove the thread from the reactor, and then
call spdk_reactor_schedule_thread().

The next patch will make this API available through JSON RPC.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I656872d32dbb469ae70f771cd0419a77236bfe18