Update patch set 9
Patch Set 9:

> Patch Set 9: Code-Review-1
> 
> Marking -1 just because I'd like to discuss this further before we merge.
> 
> We are planning to add functionality to translate poller function pointers to function names.  This would mean we would need some RPC like thread_get_pollers.  If we add that RPC, then we would not need to add poller counts to thread_get_stats.
> 
> Maybe it's OK to add this information also in thread_get_stats, but I'd like to discuss further.

Hi Jim,

This was convenient to develop the patch series but is no critical. If I skip this, could you review the remaining series?

The next patch in the series to add io channel count. I need at least an API to know if thread->io_channels is empty or not.

Will you give me your idea to the next patch too?

-Shuhei

Patch-set: 9